requirements.txt
Django==3.2
django-cte==1.1.5
django-extensions==3.1.5
djangorestframework==3.13.1